data-generator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oilgas-data-generator"
version = "1.0.0"
description = "Simulated Oil & Gas sensor data generator for NiFi monitoring platform"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (require Docker stack)",
]
addopts = "-v --tb=short"
